Drop the release milestone GitHub action
Such action exists in provider extensions repos and is mainly used to prevent Gardener maintainers to merge PRs when the extension is in release validation.
I doubt that for this repo we will have such cases as Gardener maintainers currently don't have merge rights and I don't expect them to actively create PRs against this repo.